backend/requirements.txt
asgiref==3.7.2
attrs==24.2.0
Django==4.2.16
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
ruff==0.4.5
sqlparse==0.5.0
uritemplate==4.1.1
